+++ src/main/java/bioast/mods/gt6scan/network/ScanRequestToServer.java
@@ -1,6 +1,5 @@
 package bioast.mods.gt6scan.network;
 
-import bioast.mods.gt6scan.item.ScanMode;
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import io.netty.buffer.ByteBuf;
